Tests/LibWeb/Text/expected/css/modal-dialog-display-animation-lays-out.txt

modal: true
computed display: block
has box: true
width: 698
